IDをintegerに変更したことによるrepositoryの修正

+++ src/main/java/org/actlab/msat/repository/IncomingServerDomainRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IncomingServerDomainRepository extends JpaRepository<IncomingServerDomain, Long>{
+public interface IncomingServerDomainRepository extends JpaRepository<IncomingServerDomain, Integer>{
 }
