+++ src/main/java/org/actlab/msat/repository/IncomingServerSettingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IncomingServerSettingRepository extends JpaRepository<IncomingServerSetting, Long> {
+public interface IncomingServerSettingRepository extends JpaRepository<IncomingServerSetting, Integer> {
 }
