+++ src/main/java/org/actlab/msat/repository/OutgoingServerDomainRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OutgoingServerDomainRepository extends JpaRepository<OutgoingServerDomain, Long> {
+public interface OutgoingServerDomainRepository extends JpaRepository<OutgoingServerDomain, Integer> {
 }
